api/requirements.txt
alembic==1.8.0
aniso8601==9.0.1
base58==2.1.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
paho-mqtt==1.6.1
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.37
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
gunicorn==20.1.0
flask-admin==1.6.0
flask-login==0.6.1